Move abstract extensions into PrestaShopBundle/Form/Extension
