+++ app/config/config.yml
@@ -94,8 +94,8 @@ twig:
   globals:
     webpack_server: false
     multistore_field_prefix: !php/const PrestaShopBundle\Service\Form\MultistoreCheckboxEnabler::MULTISTORE_FIELD_PREFIX
-    modify_all_shops_prefix: !php/const PrestaShopBundle\Form\Admin\Extension\ModifyAllShopsExtension::MODIFY_ALL_SHOPS_PREFIX
-    disabling_switch_prefix: !php/const PrestaShopBundle\Form\Admin\Extension\DisablingSwitchExtension::FIELD_PREFIX
+    modify_all_shops_prefix: !php/const PrestaShopBundle\Form\Extension\ModifyAllShopsExtension::MODIFY_ALL_SHOPS_PREFIX
+    disabling_switch_prefix: !php/const PrestaShopBundle\Form\Extension\DisablingSwitchExtension::FIELD_PREFIX
 
 # Swiftmailer Configuration
 swiftmailer:
