+++ src/PrestaShopBundle/Form/Extension/ColumnsNumberExtension.php
@@ -23,9 +23,10 @@
  * @copyright Since 2007 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
+
 declare(strict_types=1);
 
-namespace PrestaShopBundle\Form\Admin\Extension;
+namespace PrestaShopBundle\Form\Extension;
 
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
@@ -34,9 +35,16 @@ use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * Class HintTextExtension extends every form type with additional hint text options.
+ * Adds the "columns_number" option to all Form Types.
+ *
+ * You can use it together with the UI kit form theme to adapt the display of a form group into columns,
+ * the form theme will add a class that affects the display into flex container with fixed size for sub elements.
+ *
+ * ```
+ * 'columns_number' => 4,
+ * ```
  */
-class HintTextExtension extends AbstractTypeExtension
+class ColumnsNumberExtension extends AbstractTypeExtension
 {
     /**
      * {@inheritdoc}
@@ -45,9 +53,11 @@ class HintTextExtension extends AbstractTypeExtension
     {
         $resolver
             ->setDefaults([
-                'hint' => null,
+                'columns_number' => null,
+                'column_breaker' => false,
             ])
-            ->setAllowedTypes('hint', ['null', 'string'])
+            ->setAllowedTypes('columns_number', ['null', 'int'])
+            ->setAllowedTypes('column_breaker', ['bool'])
         ;
     }
 
@@ -56,7 +66,10 @@ class HintTextExtension extends AbstractTypeExtension
      */
     public function buildView(FormView $view, FormInterface $form, array $options)
     {
-        $view->vars['hint'] = isset($options['hint']) ? $options['hint'] : null;
+        if (!empty($options['columns_number'])) {
+            $view->vars['columns_number'] = $options['columns_number'];
+        }
+        $view->vars['column_breaker'] = $options['column_breaker'];
     }
 
     /**
