+++ src/PrestaShopBundle/Form/Extension/DownloadFileExtension.php
@@ -26,55 +26,35 @@
 
 declare(strict_types=1);
 
-namespace PrestaShopBundle\Form\Admin\Extension;
+namespace PrestaShopBundle\Form\Extension;
 
 use Symfony\Component\Form\AbstractTypeExtension;
-use Symfony\Component\Form\Extension\Core\Type\FormType;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-/**
- * Adds the "alert_message" option to all Form Types.
- *
- * You can use it together with the UI kit form theme to add alter message to your field
- */
-class AlertExtension extends AbstractTypeExtension
+class DownloadFileExtension extends AbstractTypeExtension
 {
     /**
      * {@inheritdoc}
      */
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver
             ->setDefaults([
-                'alert_message' => null,
-                'alert_type' => 'info',
-                'alert_position' => 'append',
-                'alert_title' => null,
+                'download_url' => null,
             ])
-            ->setAllowedTypes('alert_message', ['null', 'string', 'array'])
-            ->setAllowedTypes('alert_title', ['string', 'null'])
-            ->setAllowedTypes('alert_type', ['string'])
-            ->setAllowedTypes('alert_position', ['string'])
-            ->setAllowedValues('alert_position', ['append', 'prepend'])
+            ->setAllowedTypes('download_url', ['null', 'string'])
         ;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function buildView(FormView $view, FormInterface $form, array $options)
+    public function buildView(FormView $view, FormInterface $form, array $options): void
     {
-        if (!empty($options['alert_message'])) {
-            $view->vars['alert_message'] = is_string($options['alert_message']) ? [$options['alert_message']] : $options['alert_message'];
-            $view->vars['alert_type'] = $options['alert_type'];
-            $view->vars['alert_position'] = $options['alert_position'];
-
-            if (is_string($options['alert_title'])) {
-                $view->vars['alert_title'] = $options['alert_title'];
-            }
-        }
+        $view->vars['download_url'] = $options['download_url'];
     }
 
     /**
@@ -82,6 +62,6 @@ class AlertExtension extends AbstractTypeExtension
      */
     public static function getExtendedTypes(): iterable
     {
-        return [FormType::class];
+        return [FileType::class];
     }
 }
