+++ src/PrestaShopBundle/Form/Extension/ExternalLinkExtension.php
@@ -24,18 +24,30 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShopBundle\Form\Admin\Extension;
+declare(strict_types=1);
+
+namespace PrestaShopBundle\Form\Extension;
 
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
+use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * Class HelpTextExtension extends every form type with additional help text options.
+ * Adds the "external_link" option to all Form Types.
+ *
+ * You can use it together with the UI kit form theme to add external links:
+ *
+ * ```
+ * 'external_link' => [
+ *   'link' => 'foo bar',
+ *   'text' => 'foo bar',
+ * ],
+ * ```
  */
-class HelpTextExtension extends AbstractTypeExtension
+class ExternalLinkExtension extends AbstractTypeExtension
 {
     /**
      * {@inheritdoc}
@@ -44,9 +56,18 @@ class HelpTextExtension extends AbstractTypeExtension
     {
         $resolver
             ->setDefaults([
-                'help' => null,
+                'external_link' => null,
             ])
-            ->setAllowedTypes('help', ['null', 'string'])
+            ->setAllowedTypes('external_link', ['null', 'array'])
+            ->setNormalizer('external_link', function (Options $options, $value) {
+                if (null === $value) {
+                    return null;
+                }
+
+                $resolver = $this->getExternalLinkResolver();
+
+                return $resolver->resolve($value);
+            })
         ;
     }
 
@@ -55,7 +76,9 @@ class HelpTextExtension extends AbstractTypeExtension
      */
     public function buildView(FormView $view, FormInterface $form, array $options)
     {
-        $view->vars['help'] = $options['help'] ?? null;
+        if (!empty($options['external_link'])) {
+            $view->vars['external_link'] = $options['external_link'];
+        }
     }
 
     /**
@@ -64,5 +87,29 @@ class HelpTextExtension extends AbstractTypeExtension
     public static function getExtendedTypes(): iterable
     {
         return [FormType::class];
+    }
+
+    /**
+     * @return OptionsResolver
+     */
+    private function getExternalLinkResolver(): OptionsResolver
+    {
+        $externalLinkResolver = new OptionsResolver();
+        $externalLinkResolver
+            ->setRequired(['href', 'text'])
+            ->setDefaults([
+                'attr' => [],
+                'align' => 'left',
+                'position' => 'append',
+            ])
+            ->setAllowedTypes('href', 'string')
+            ->setAllowedTypes('text', 'string')
+            ->setAllowedTypes('align', 'string')
+            ->setAllowedTypes('position', 'string')
+            ->setAllowedTypes('attr', ['null', 'array'])
+            ->setAllowedValues('position', ['append', 'prepend'])
+        ;
+
+        return $externalLinkResolver;
     }
 }
