+++ src/PrestaShopBundle/Form/Extension/HintTextExtension.php
@@ -23,8 +23,9 @@
  * @copyright Since 2007 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
+declare(strict_types=1);
 
-namespace PrestaShopBundle\Form\Admin\Extension;
+namespace PrestaShopBundle\Form\Extension;
 
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
@@ -33,17 +34,9 @@ use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * Adds the "row_attr" option to all Form Types.
- *
- * You can use it together with the UI kit form theme to add classes to your from rows:
- *
- * ```
- * 'row_attr' => [
- *   'class' => 'foo bar',
- * ],
- * ```
+ * Class HintTextExtension extends every form type with additional hint text options.
  */
-class RowAttributesExtension extends AbstractTypeExtension
+class HintTextExtension extends AbstractTypeExtension
 {
     /**
      * {@inheritdoc}
@@ -52,9 +45,9 @@ class RowAttributesExtension extends AbstractTypeExtension
     {
         $resolver
             ->setDefaults([
-                'row_attr' => null,
+                'hint' => null,
             ])
-            ->setAllowedTypes('row_attr', ['null', 'array'])
+            ->setAllowedTypes('hint', ['null', 'string'])
         ;
     }
 
@@ -63,7 +56,7 @@ class RowAttributesExtension extends AbstractTypeExtension
      */
     public function buildView(FormView $view, FormInterface $form, array $options)
     {
-        $view->vars['row_attr'] = $options['row_attr'] ?? [];
+        $view->vars['hint'] = isset($options['hint']) ? $options['hint'] : null;
     }
 
     /**
