+++ src/PrestaShopBundle/Form/Extension/MultistoreExtension.php
@@ -26,42 +26,48 @@
 
 declare(strict_types=1);
 
-namespace PrestaShopBundle\Form\Admin\Extension;
+namespace PrestaShopBundle\Form\Extension;
 
 use Symfony\Component\Form\AbstractTypeExtension;
-use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class DownloadFileExtension extends AbstractTypeExtension
+class MultistoreExtension extends AbstractTypeExtension
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver
-            ->setDefaults([
-                'download_url' => null,
-            ])
-            ->setAllowedTypes('download_url', ['null', 'string'])
-        ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function buildView(FormView $view, FormInterface $form, array $options): void
-    {
-        $view->vars['download_url'] = $options['download_url'];
-    }
-
     /**
      * {@inheritdoc}
      */
     public static function getExtendedTypes(): iterable
     {
-        return [FileType::class];
+        return [FormType::class];
+    }
+
+    /**
+     * @param FormView $view
+     * @param FormInterface $form
+     * @param array $options
+     */
+    public function buildView(FormView $view, FormInterface $form, array $options): void
+    {
+        parent::buildView($view, $form, $options);
+
+        $view->vars = array_replace($view->vars, ['multistore_dropdown' => $options['multistore_dropdown']]);
+    }
+
+    /**
+     * @param OptionsResolver $resolver
+     */
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        parent::configureOptions($resolver);
+
+        $resolver->setDefaults(
+            [
+                'multistore_dropdown' => false,
+                'multistore_configuration_key' => null,
+            ]
+        );
     }
 }
