+++ src/PrestaShopBundle/Resources/config/services/core/form/data_formatter.yml
@@ -2,7 +2,7 @@ services:
   _defaults:
     autowire: true
     bind:
-      $modifyAllNamePrefix: !php/const PrestaShopBundle\Form\Admin\Extension\ModifyAllShopsExtension::MODIFY_ALL_SHOPS_PREFIX
+      $modifyAllNamePrefix: !php/const PrestaShopBundle\Form\Extension\ModifyAllShopsExtension::MODIFY_ALL_SHOPS_PREFIX
 
   PrestaShop\PrestaShop\Core\Form\IdentifiableObject\DataFormatter\BulkCombinationFormDataFormatter: ~
